remove faucetfaucet
